+++ Fabric/src/main/java/fuzs/forgeconfigapiport/impl/config/ForgeConfigRegistryV2Impl.java
@@ -4,7 +4,7 @@ import fuzs.forgeconfigapiport.api.config.v2.ForgeConfigRegistry;
 import net.minecraftforge.fml.config.IConfigSpec;
 import net.minecraftforge.fml.config.ModConfig;
 
-public final class ForgeConfigRegistryImpl implements ForgeConfigRegistry {
+public final class ForgeConfigRegistryV2Impl implements ForgeConfigRegistry {
 
     @Override
     public ModConfig register(String modId, ModConfig.Type type, IConfigSpec<?> spec) {