+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v9.1.0-1.20] - 2023-11-30
+### Added
+- Added support for the `net.neoforged` namespace in addition to `net.minecraftforge`
+- Support for Minecraft Forge will be dropped in the long term
+- At the moment config systems from both mod loaders are fully supported, except the `/config` command which now only works for NeoForge configs
+
 ## [v9.0.0-1.20] - 2023-09-22
 - Ported to Minecraft 1.20.2